requirements.txt
aniso8601==9.0.1
attrs==21.4.0
bcrypt==3.2.2
blinker==1.4
cffi==1.15.0
click==8.1.3
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.1
Flask-Mail==0.9.1
Flask-RESTful==0.3.9
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.6.0
MarkupSafe==2.1.1
pycparser==2.21
PyJWT==2.4.0
pyrsistent==0.18.1
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.37
Werkzeug==2.1.2
zipp==3.8.0
